requirements.txt
fastapi==0.95.0
uvicorn==0.21.1
pydantic==1.10.7
torch==2.0.0
transformers==4.27.4
pandas==2.0.0
numpy==1.24.2
scikit-learn==1.2.2
seqeval==1.2.2
apscheduler==3.10.1
python-multipart==0.0.6
asyncpg==0.30.0
python-dotenv==1.1.0
pytorch-crf==0.7.2
